# frontend/openapi.yaml
openapi: 3.0.0
info:
  title: NitiNirmaan API
  description: API contract for the NitiNirmaan frontend application.
  version: 1.0.0
servers:
  - url: http://localhost:3000/api
    description: Local development server
  - url: https://api.nitinirmaan.org/v1
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
        avatarUrl:
          type: string
        level:
          type: object
          properties:
            levelNumber:
              type: integer
            title:
              type: string
              example: Strategist
        organization:
          type: string
    Project:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        status:
          type: string
          enum: [Draft, Simulation Failed, Ready for Export, In Progress]
        editedAt:
          type: string
          format: date-time
        logicHealth:
          type: integer
          minimum: 0
          maximum: 100
        state:
          type: string
        district:
          type: string
        domain:
          type: string
          enum: [FLN, Career Readiness, STEM, Other]
        outcome:
          type: string
        aiCompanion:
          type: string
          enum: [supportive, critical]
    Pattern:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        institution:
          type: string
        forkedCount:
          type: integer
        successRate:
          type: string
          enum: [High, Moderate, Low]
        theme:
          type: array
          items:
            type: string
        geography:
          type: array
          items:
            type: string
        scale:
          type: array
          items:
            type: string
        nodes:
          type: array
          items:
            type: object
            properties:
              type:
                type: string
              label:
                type: string
              x:
                type: number
              y:
                type: number
        connections:
          type: array
          items:
            type: object
            properties:
              from:
                type: integer
              to:
                type: integer
        metadata:
          type: object
          properties:
            geography:
              type: string
            operatingScale:
              type: string
            stakeholders:
              type: array
              items:
                type: string
            constraints:
              type: array
              items:
                type: string
paths:
  /auth/login:
    post:
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
  /auth/signup:
    post:
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                orgName:
                  type: string
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
  /auth/google:
    post:
      summary: Login with Google
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
      responses:
        '200':
          description: Successful Google login
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
  /user/me:
    get:
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /projects:
    get:
      summary: List all projects (missions) for the user
      security:
        - bearerAuth: []
      parameters:
        - name: sort
          in: query
          schema:
            type: string
            enum: [recent, name, health]
      responses:
        '200':
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Project'
    post:
      summary: Create a new project (mission)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - projectName
                - domain
                - outcome
              properties:
                projectName:
                  type: string
                state:
                  type: string
                district:
                  type: string
                domain:
                  type: string
                outcome:
                  type: string
                aiCompanion:
                  type: string
                  enum: [supportive, critical]
      responses:
        '201':
          description: Project created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
  /projects/{id}:
    get:
      summary: Get project details by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Project details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
    put:
      summary: Update project details
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Project'
      responses:
        '200':
          description: Project updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
    delete:
      summary: Delete or archive a project
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Project deleted
  /projects/{id}/duplicate:
    post:
      summary: Duplicate an existing project
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '201':
          description: Project duplicated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
  /patterns:
    get:
      summary: List pattern templates
      security:
        - bearerAuth: []
      parameters:
        - name: theme
          in: query
          schema:
            type: string
        - name: geography
          in: query
          schema:
            type: string
        - name: scale
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of templates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Pattern'
  /patterns/{id}/fork:
    post:
      summary: Fork a pattern to create a new project
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '201':
          description: Pattern forked into new project
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'